feat(ci): add GitHub Actions workflow for Docker image build and push
Introduce a new GitHub Actions workflow to automate the building and pushing of Docker images. This workflow supports specifying Node.js version and image tag version via workflow dispatch inputs, enhancing flexibility and automation in the CI/CD pipeline.

fix(ui): update flowise-embed dependency to specific GitHub branch

Change the `flowise-embed` dependency in `package.json` to point to a specific branch on GitHub. This ensures that the project uses a stable and specific version of the dependency, avoiding potential issues with the 'latest' tag.

refactor(ui): update branding from Flowise to Memox

Update the HTML metadata and title to reflect the new branding from Flowise to Memox. This includes changes to the title, description, and various meta tags to ensure consistency with the new brand identity.